Correctly calculate roots of PG before submission
The removal of the last node in the PG seems to correspond to the logic
of that last node having the reproducibility data of the graph instead
of an actual drop specification (this is the implication I gathered by
looking at the rest of the changes in 58138a30b, the commit where this
change was introduced). This reproducibility data is optional though,
and therefore the "dlg submit" tool, via this submit() function, was
broken and removed the last drop specification from the calculation of
the graph roots, which in turn had the effect that the session didn't
trigger one of the roots of the graph at graph submission time, leaving
the graph in Running state forever.

This change fixes the submission of "normal" graphs (i.e., without the
extra reproducibility data object at the end), making the tool usable
for existing graphs. This will clearly break graphs *with* the
reproducibility data, but those are still fewer in number.

I found this problem while working on YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>