Move all logging to lazy % evaluation
The current daliuge code base had a mixture of logging statements that
were lazily and eagerly formatting their messages. On top of that,
different formatting options were being used (f-strings, .format() and
%-based formatting).

This commit unifies all the statements to use the same convention. Lazy
evaluation is the preferred way to avoid performance penalties (see
985ca8f for an example). %-based formatting is the way the logging
library does lazy interpolation internally, and while alternatives are
supported they're always trickier to use.

This is one of the main points to address in YAN-981.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>