Add initial pylint configuration file
Initially pylint is configured to *not* run most checks, enabling only
those that we want to tackle. Now that we have moved all logging
statements to a unified convention we can use pylint to check this,
which in turn can be added as a step in the CI pipeline to enforce it.

This work is part of YAN-981.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>