Store oids in set for faster lookups
This function first collects all OIDs, and then checks different OIDs
for existence in that collection. We have been using a list to store all
OIDs though, making the lookups not as fast as they could have been.

This commit changes the collection holding all OIDs from a list to a
set. This speeds up the lookups of further OIDs in the rest of the
method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>