Lessen resource usage when multiprocessing not in use
One of the problems with the current multiprocessing support in daliuge
is that it requires too many resources. This is because each DropProxy
object is creating its own RPCClient, which internally create their own
ZMQ context. This not only increases the runtime requirements of the
system (more objects need to be created, etc.) but also puts pressure on
system resources. This situation is described in detail in YAN-976, and the
test/manager/test_dm.py::TestDM::test_many_relationships unit test in
particular triggers this issue with ease.

This commit implements a partial fix for this issue. While it doesn't
take care of the problem itself, it identifies when a DropProxy is
created from a NodeManager that is *not* using multiprocessing support,
and avoid creating an RPCClient instance itself, instead using the
client provided as a parameter. This brings stability back to the
daliuge runtime when multiprocessing isn't used, containing the resource
over-usage to only when strictly necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>