Add note about temporary fix in get_roots
Graphs are not formally specified, so we now have ended up with graphs
that do not contain drops only, and that have customisations to which
the rest of the system needs to adjust. This note here (together with a
new tech debt ticket) should help in clearing out the situation in the
future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>