Skip multiprocessing tests by default
These tests are not stable (see YAN-976 for some examples), but their
instability has been hidden by the fact that they are never run on
Travis anyway. On my laptop though they often cause issues, making the
whole test suite unreliable.

This commit changes the skip condition for the multiprocessing tests to
only run if a specific environment variable is set (it isn't by
default). This will allow us to continue hiding the problem, but in a
more controller manner, while also giving us an easy switch to turn the
tests on when we start addressing their problems.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>