Load and use JSON schema definition always
The JSON schema as previously stored at the root directory of the
daliuge-translator package, making it unavailable to the package
runtime, and therefore limiting its usage to deployments where the file
would need to be manually copied to a known location.

This limitation is unnecessary if the file is stored as part of the
daliuge-translator package data. This commit implements that, moving
(and renaming) the file under dlg/dropmake/lg.graph.schema, declaring it
as part of the package data, and loading its contents unconditionally at
package import time. This means we can now run LG schema validation in
all cases, including during unit testing.

This work is part of YAN-899.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>