Fix NodeManager MRO and start/shutdown calls
The NodeManager hierarchy is fairly complex, and some of its details
were a bit twisted. One of them was the sequence of the start/shutdown
calls, which in their previous form did not allow for the methods to be
overloaded by the NodeManagerBase class itself, and couldn't participate
on the MRO resolution used by the super() calls.

This commit re-organises the MRO of the NodeManager class by placing
NodeManagerBase as its first child. This allows it to have functions
that can then call the rest of the chain. This in turn meant we had to
remove a few of the methods that were marked as "abstract", but that's
not a problem as it's mostly dead code.

A second change done by this class is that the start() call is issued
from the NodeManager class rather than from NodeManagerBase. The latter
was not the best place to issue this call because it wasn't the final
class if the hierarchy, as users instantiate NodeManager instead. This
change also allows NodeManager to create its ZeroMQ context object
within its own start() method rather than in the __init__ method, which
wasn't aligned with the rest of the classes in the package.

With this change we can now correctly override the start() method in
NodeManagerBase, which will be needed to start the DLM background tasks.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>