Manage ZeroRPCClient's own context correctly
The ZeroRPCClient class usually is instantiated as part of the
NodeManager class, in which case there already is a ZeroMQ context
crated for it to use. There are other cases however where the
ZeroRPCClient class is used own its own (e.g., by the off-process dynlib
support), and in these situations the class creates a ZeroMQ context for
itself.

There were some small problems with how this own context was created and
managed though. Firstly, the context was created in the __init__
function rather than in the start function. While this is not a *real*
problem, it's not aligned with how the rest of the RPC classes operate.
Secondly, and more importantly, this context was never destroyed
explicitly, making its destruction non-deterministic and up to how the
system operates. Until that destruction happened, system resources would
have been unnecessarily kept held.

This commit moves the creation of ZeroRPCClient's own context, when
needed, to the start method rather than the __init__ method, to align it
with the rest of the classes. A new boolean member stores whether this
context has been created by ZeroRPCClient itself or not. Finally, in
the stop method, contexts created by ZeroRPCClient are terminated.