Cache class members calculation on drop creation
When a drop is created, its class member attributes are inspected to
perform automatic extraction of keyword arguments, if needed. This
inspection was taking place every time a drop was created though, even
though different drops of the same type would yield the same members.

This commit adds a cache to the calculation of class member attributes
done at drop instantiation time. The cache is filled for each new class
that is inspected, and its results are reused for successive drops of
types that have been previously inspected.

This work is part of LIU-268.