Add initial version of GitHub Actions
There are four workflows defined in our new set of GitHub Actions that
mimic the tasks we currently perform in Travis: run unit tests, build
component palettes, do linting, and build the documentation. Not only do
the workflows mimic the actions performed in Travis, but they also use
secrets to do authenticated actions, like pushing commits to the EAGLE
test repository, or upload data to coveralls.

This work is part of YAN-1041.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>