Run lgweb tests on a different port
The tests for the lgweb module start the lgweb server on port 8084. This
has never been a problem until now: in GitHub Actions there is a process
that has port 8084 already bound, making our lgweb server fail to start,
which in turn makes our daliuge-translator tests fail.

This commit sets a different port for the lgweb tests to run on. With
this, our GitHub Actions CI can run our tests to completion.

This work is part of YAN-1041.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>