Don't keep a reference to funcargs and pargs
These two variables are used only in the run() method of PyFuncApp, and
thus do not need to be kept around in self. Doing so prevents their
contents from being freed, which keeps memory usage high if a reference
to the PyFuncApp instance is still around.

This work is part of YAN-1048.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>