Free DLM's drop references on session destruction
When a session is destroyed all its drops are meant to disappear from
memory, and thus the data they hold should also be freed. This however
is not necessarily the case when the DLM is in use, as it will hold
references to these drops even after a session is destroyed.

This commit adds the necessary changes to remove references from the DLM
to drops in a session that is about to be destroyed. This should allow
those objects to be destroyed once the session is actually destroyed,
which in turn should free the underlying resources they hold.

This work is part of YAN-1048.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>