LIU-308: Ensure drop precious flag is still handled
To ensure backwards compatibility rename it to persist when constructing
the drop instances. We also could add duplicated logic to do this during
translation from LG->PG, however we still need to handle this here as
old PGs might still have the flag. Additionally there don't seem to be
any obvious locations in the translator to add this logic as it is
exclusively dealing with graph traversal and just passes through any
properties is doesn't deal with straight through to the PG.