Add default value for otherwise empty list
The optparse module documentation states that an empty list is created
"when [...] first encounters this option on the command line", meaning
that if no option is given then no empty list is created. This was
causing a TypeError afterwards because the destination object (a None
when no algorithm parameters were passed) was being iterated over.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>