Introduce new ProxyInfo object
This object is a simple dataclass that holds the information that
DropProxy objects need to perform their duties. This is the *static*
information -- the DropProxy still requires a RPCClient to actually
interact with a remote RPC server.

This little utility class encapsulates some of the behavior that was
previously found in the dynlib module, where we setup proxies for
newly-created processes. While this isn't a great change in itself, it
prepares the codebase for a much bigger change: the introduction of
subprocesses for the effective execution of drop apps.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>