Delegate execution of AppDROP.run() to DropRunner
By extracting this into an ABC we can change the implementation (by
default it will occur synchronously, but in the context of a
NodeManager, it will execute on a ThreadPool). This sets the stage for
an implementation that utilises Processes for true parrallelisation.

If a drop is asynchronously executed a seperate, daemon thread is
created to wait for execution to finish. This job shouldn't be run on
the same pool as the DropRunner as it is easy to deadlock by running out
of executors in the implementing pool (e.g. a ThreadPool with
max_executors=1, async_execute() is submitted, takes up the single
Thread prior to the actual run() method being submitted to the pool. In
this case run() will never execute as there is never a free thread and
async_execute() blocks forever). Note that the daemon thread will not be
terminated until the process exits, so this causes a memory leak and
will need to be addressed in future.