Store session ID instead of full session in drops
Drops stored a reference to the session that contained them in the
_dlg_session attribute. This attribute was then later used to extract
the ID of such session via *its* sessionID attribute. The existence of
this ID is then seen in a number of places across the code, causing a
number of effects (setting environment variables, logging the session
ID, etc).

A closer inspection to the code revealed that the sessionID attribute
was the only attribute ever read from the drop._dlg_session attribute.
Thus, storing the full session object is unnecessary. While during
normal usage this doesn't matter much, there is a negative effect on
serialisation of drops, which cannot be achieved because the Session
object bound to one of its attributes isn't serialisable (it holds not
only a lock, but also a reference to the Node Manager, which contains
open file descriptors, thread pools, and more).

This commit removes the internal _dlg_session attribute from the
AbstractDROP class, and replaces it with a _dlg_session_id. To make
things easier overall we default its value to an empty string, both when
the drops are created directly (e.g., MyDrop()) and when they are
constructed from the graph_loader module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>