LIU-398: PyFuncScatterApp (WIP)
Start experimenting with creating a generic PyFuncScatterApp. The intention is we wrap a user-specified python function with the existing GenericScatterApp code, so we don't have to have users create entirely new Apps; all they have to do is run dlg_paletteGen on their method, and plug-in the right input parameters.