Add another small bump, and update strategy to no longer run fail-fast.
Fail-fast is normally 'true' by default and is useful in matrix strategies to ensure that large/long jobs which are all likely to fail will stop the minute on of the jobs indicates there's a bug/failure. However, we have smaller job numbers and often one fails due to intermittent performance issues which are not noticeable for a different job (i.e. Python version number).