LIU-5: Add input-output FileDROP behaviour with prefix and suffix support.
An existing limitation of the DALiuGE engine is one of it's strengths; we create our own path names for intermediary data products because hard-coding them is prone to error, and so we have complete knowledge of the data.

The problem here when we have applications or functions that have an input parameter that is the name of the file (i.e. a runtime-generated output DROP). To fix this, I have:

- Added support for an input-output flag for a given parameter that is an ApplicationArgument
- Provided a way to reference which output the input-output parameter we want to associated the flagged parameter
- Provided support for prefix and suffix behaviour