LIU-3: Add preliminary support for global variable construct
I've created a "GlobalVariables" construct that acts as our wrapper for any variable that would be used multiple times across the graph. I think a Construct is the appropriate type for this, as we will remove this in the translator (like the other Constructs) and therefore it does not end up becoming an actual DROP. It's also simpler because we can have no fields beyond whatever parameters the user has put into it.

I have tested this with a super basic example; I'll add tests in a future commit.

Signed-off-by: Ryan Bunney <ryanbunney@fastmail.com>