LIU-517: Fix node_str addition
This issue was introduced with the Node class back in https://github.com/ICRAR/daliuge/pull/283. The purpose of that class was to improve the wrapping of DRopManager ports so we aren't reliant on default ports across the nodes, which becomes a problem if you want to run more than one manager on a single machine (port conflicts incur). One side effect of this was updating the parameters to some of our REST functions from `node` to `node_str`. This was intended to make it clear that we are getting a 'string' node rather than a Node node, but it actually broke out API calls scattered around our web interface. I've corrected the mistake and a couple of other outstanding issues.

Signed-off-by: Ryan Bunney <ryanbunney@fastmail.com>