LIU-451: Convert DirectoryContainer to DirectoryDROP
DirectoryDROP is going to be a proxy for directories on the file system.

This commit will cause a bunch of tests to fail as I have not reorganised the existing unittests, but we can successfully access the directory using path attribute and interact with things like the CopyApp and BashShellAppDROPs.

Signed-off-by: Ryan Bunney <ryanbunney@fastmail.com>