LIU-530: Address issue with branch not working on multiple nodes.
The cause of this is long and complex, but the crux of the matter is we remove 'inputs' from some drops in the partitioning between NodeManagers, which causes them not to be tracked properly.

This commit is currently quite an ugly use of data we already have; the root cause is a poor data model and the separation of inputs and ports. This needs further work and is not appropriate for a 'patch', as will require some significant re-design and refactoring of both the translator and engine. For the purposes of getting things running, we are using the outputPorts that were previously added for the Branch in the translator, and which get sucked in the AbstractDROP initialisation because we store everything in the dropSpec in parameters (because we are lazy).

Signed-off-by: Ryan Bunney <ryanbunney@fastmail.com>